requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Telegram scraping
telethon==1.34.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# dbt
dbt-core==1.7.0
dbt-postgres==1.7.0

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# YOLO
ultralytics==8.1.0
opencv-python==4.8.1.78

# Orchestration
dagster==1.5.0
dagster-webserver==1.5.0

# Utilities
pandas==2.1.3
numpy==1.26.2
pillow==10.1.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
